requirements.txt
appnope==0.1.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cloudstorage==0.10.1
cycler==0.10.0
decorator==4.4.2
google-api-core==1.23.0
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-cloud-vision==2.0.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
inflection==0.5.1
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kiwisolver==1.2.0
libcst==0.3.13
matplotlib==3.3.2
mypy-extensions==0.4.3
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
PySocks==1.7.1
pytesseract==0.3.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-magic==0.4.18
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
traitlets==5.0.5
tweepy==3.9.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
wcwidth==0.2.5
webapp2==2.5.2
